[*] merge cpp together
