+++ lib/types/objects/Boolean/methods.cpp
