+++ lib/types/objects/Number/methods.cpp
