+++ lib/types/objects/Object/methods.cpp
@@ -1,10 +1,117 @@
-#include "internals/Object.hpp"
 #include "types/objects/Object/JsObject.hpp"
-#include "types/objects/Function/JsFunction.hpp"
 #include "utils/Compare.hpp"
 #include "utils/Convert.hpp"
 #include "utils/Is.hpp"
 
+// prototype methods
+JS::Any JS::Object::toString(const std::vector<JS::Any>& args) {
+    // TODO Implement correctly ToObject to make all the methods work (ToObject should always return an object so the
+    // compare type should always be true)
+    switch (args[0].getValue().index()) {
+        case JS::UNDEFINED:
+            return JS::Any("[object Undefined]");
+        case JS::NULL_TYPE:
+            return JS::Any("[object Null]");
+    }
+    JS::Any O = JS::CONVERT::ToObject(args[0]);
+    // TODO to remove when ToObject is implemented correctly
+    if (!JS::COMPARE::Type(O, JS::OBJECT)) {
+        throw std::runtime_error(
+            "TypeError: Object.prototype.toString called on non-object (this should never happen)"); // this should
+                                                                                                     // never happen
+    }
+    return JS::Any("[object " + std::get<std::shared_ptr<JS::InternalObject>>(O.getValue())->class_name + "]");
+}
+
+JS::Any JS::Object::toLocaleString(const std::vector<JS::Any>& args) {
+    JS::Any O = JS::CONVERT::ToObject(args[0]);
+    // TODO to remove when ToObject is implemented correctly
+    if (!JS::COMPARE::Type(O, JS::OBJECT)) {
+        throw std::runtime_error(
+            "TypeError: Object.prototype.toLocaleString called on non-object (this should never happen)"); // this
+                                                                                                           // should
+                                                                                                           // never
+                                                                                                           // happen
+    }
+    JS::Any toString = std::get<std::shared_ptr<JS::InternalObject>>(O.getValue())->get("toString");
+    if (JS::IS::Callable(toString)) {
+        return toString();
+    }
+    throw std::runtime_error("TypeError: Object.prototype.toLocaleString called on non-object");
+}
+
+JS::Any JS::Object::valueOf(const std::vector<JS::Any>& args) {
+    JS::Any O = JS::CONVERT::ToObject(args[0]);
+    // TODO to remove when ToObject is implemented correctly
+    if (!JS::COMPARE::Type(O, JS::OBJECT)) {
+        throw std::runtime_error(
+            "TypeError: Object.prototype.valueOf called on non-object (this should never happen)"); // this should never
+                                                                                                    // happen
+    }
+    return O;
+}
+
+JS::Any JS::Object::hasOwnProperty(const std::vector<JS::Any>& args) {
+    std::string P = JS::CONVERT::ToString(args[1]);
+    JS::Any O = JS::CONVERT::ToObject(args[0]);
+    // TODO to remove when ToObject is implemented correctly
+    if (!JS::COMPARE::Type(O, JS::OBJECT)) {
+        throw std::runtime_error(
+            "TypeError: Object.prototype.hasOwnProperty called on non-object (this should never happen)"); // this
+                                                                                                           // should
+                                                                                                           // never
+                                                                                                           // happen
+    }
+    std::shared_ptr<JS::InternalObject> obj = std::get<std::shared_ptr<JS::InternalObject>>(O.getValue());
+    if (obj->getOwnProperty(P).has_value()) {
+        return JS::Any(true);
+    }
+    return JS::Any(false);
+}
+
+JS::Any JS::Object::isPrototypeOf(const std::vector<JS::Any>& args) {
+    if (args.empty() || args[1].getValue().index() != JS::OBJECT) {
+        return JS::Any(false);
+    }
+    JS::Any O = JS::CONVERT::ToObject(args[0]);
+    // TODO to remove when ToObject is implemented correctly
+    if (!JS::COMPARE::Type(O, JS::OBJECT)) {
+        throw std::runtime_error(
+            "TypeError: Object.prototype.isPrototypeOf called on non-object (this should never happen)"); // this should
+                                                                                                          // never
+                                                                                                          // happen
+    }
+    std::shared_ptr<JS::InternalObject> obj = std::get<std::shared_ptr<JS::InternalObject>>(O.getValue());
+    std::shared_ptr<JS::InternalObject> V = std::get<std::shared_ptr<JS::InternalObject>>(args[1].getValue());
+    while (V->prototype != nullptr) {
+        V = V->prototype;
+        if (V.get() == obj.get()) {
+            return JS::Any(true);
+        }
+    }
+    return JS::Any(false);
+}
+
+JS::Any JS::Object::propertyIsEnumerable(const std::vector<JS::Any>& args) {
+    std::string P = JS::CONVERT::ToString(args[1]);
+    JS::Any O = JS::CONVERT::ToObject(args[0]);
+    // TODO to remove when ToObject is implemented correctly
+    if (!JS::COMPARE::Type(O, JS::OBJECT)) {
+        throw std::runtime_error("TypeError: Object.prototype.propertyIsEnumerable called on non-object (this should "
+                                 "never happen)"); // this should never happen
+    }
+    std::shared_ptr<JS::InternalObject> obj = std::get<std::shared_ptr<JS::InternalObject>>(O.getValue());
+    std::optional<JS::Attribute> desc = obj->getOwnProperty(P);
+    if (!desc.has_value()) {
+        return JS::Any(false);
+    }
+    return JS::Any(JS::IS::DataDescriptor(desc.value()) && std::get<JS::DataDescriptor>(desc.value()).enumerable ||
+                   JS::IS::AccessorDescriptor(desc.value()) &&
+                       std::get<JS::AccessorDescriptor>(desc.value())
+                           .enumerable); // TODO can be optimized with a genericDescriptor with enumerable
+}
+
+// static
 JS::Any JS::Object::getPrototypeOf(const std::vector<JS::Any>& args) {
     if (args.size() < 2 || args[1].getValue().index() != JS::OBJECT) {
         throw std::runtime_error("TypeError: Object.getPrototypeOf called on non-object"); // TODO: make this a JS error
@@ -88,7 +195,7 @@ JS::Any JS::Object::defineProperties(const std::vector<JS::Any>& args) {
         O->defineOwnProperty(key, JS::CONVERT::ToPropertyDescriptor(props->get(key)), true);
     }
     return args[0];
-} // namespace JS
+}
 
 JS::Any JS::Object::seal(const std::vector<JS::Any>& args) {
     if (args.size() < 2 || args[1].getValue().index() != JS::OBJECT) {
