--- lib/types/objects/String/static.cpp
@@ -1,5 +1,0 @@
-#include "types/objects/String/JsString.hpp"
-
-JS::Any JS::String::fromCharCode(const std::vector<JS::Any>& args) {
-    throw std::runtime_error("not implemented");
-}
